+++ examples/sinkiq.rb
@@ -12,7 +12,7 @@ class SinatraWorker
   include Sidekiq::Worker
 
   def perform(msg="lulz you forgot a msg!")
-    $redis.lpush("sinkiq-example-messages", msg)      
+    $redis.lpush("sinkiq-example-messages", msg)
   end
 end