+++ lib/sidekiq/extensions/generic_proxy.rb
@@ -13,7 +13,7 @@ module Sidekiq
         # to JSON and then deserialized on the other side back into a
         # Ruby object.
         obj = [@target, name, args]
-        ::Sidekiq::Client.push('class' => @performable.name, 'args' => [::YAML.dump(obj)])
+        @performable.perform_async(::YAML.dump(obj))
       end
     end