+++ lib/sidekiq/manager.rb
@@ -4,6 +4,7 @@ require 'multi_json'
 
 require 'sidekiq/util'
 require 'sidekiq/processor'
+require 'connection_pool/version'
 
 module Sidekiq
 
@@ -19,7 +20,8 @@ module Sidekiq
     trap_exit :processor_died
 
     def initialize(options={})
-      logger.info "Booting sidekiq #{Sidekiq::VERSION} with Redis at #{redis.client.location}"
+      logger.info "Booting sidekiq #{Sidekiq::VERSION} with Redis at #{redis {|x| x.client.location}}"
+      logger.info "Running in #{RUBY_DESCRIPTION}"
       logger.debug { options.inspect }
       @count = options[:concurrency] || 25
       @queues = options[:queues]
@@ -31,33 +33,33 @@ module Sidekiq
       @count.times.each { @ready << Processor.new_link(current_actor) }
     end
 
-    def stop
+    def stop(options={})
+      shutdown = options[:shutdown]
+      timeout = options[:timeout]
+
       @done = true
-      @ready.each(&:terminate)
+      @ready.each { |x| x.terminate if x.alive? }
       @ready.clear
 
-      redis.with_connection do |conn|
+      redis do |conn|
         workers = conn.smembers('workers')
         workers.each do |name|
           conn.srem('workers', name) if name =~ /:#{process_id}-/
         end
       end
 
-      if @busy.empty?
-        return signal(:shutdown)
-      end
-
-      logger.info("Pausing 5 seconds to allow workers to finish...")
-      after(5) do
-        @busy.each(&:terminate)
-        redis.with_connection do |conn|
-          conn.multi do
-            @busy.each do |busy|
-              conn.lpush("queue:#{busy.queue}", busy.msg)
-            end
-          end
+      if shutdown
+        if @busy.empty?
+          # after(0) needed to avoid deadlock in Celluoid after USR1 + TERM
+          return after(0) { signal(:shutdown) }
+        else
+          logger.info { "Pausing #{timeout} seconds to allow workers to finish..." }
         end
-        signal(:shutdown)
+
+        after(timeout) do
+          @busy.each { |x| x.terminate if x.alive? }
+          signal(:shutdown)
+        end
       end
     end
 
@@ -73,11 +75,11 @@ module Sidekiq
       watchdog('sidekiq processor_done crashed!') do
         @done_callback.call(processor) if @done_callback
         @busy.delete(processor)
-        processor.msg = processor.queue = nil
         if stopped?
           processor.terminate if processor.alive?
+          signal(:shutdown) if @busy.empty?
         else
-          @ready << processor
+          @ready << processor if processor.alive?
         end
         dispatch
       end
@@ -89,18 +91,18 @@ module Sidekiq
       unless stopped?
         @ready << Processor.new_link(current_actor)
         dispatch
+      else
+        signal(:shutdown) if @busy.empty?
       end
     end
 
     private
 
     def find_work(queue)
-      msg = redis.lpop("queue:#{queue}")
+      msg = redis { |x| x.lpop("queue:#{queue}") }
       if msg
         processor = @ready.pop
         @busy << processor
-        processor.msg = msg
-        processor.queue = queue
         processor.process!(MultiJson.decode(msg), queue)
       end
       !!msg
@@ -110,6 +112,10 @@ module Sidekiq
       watchdog("Fatal error in sidekiq, dispatch loop died") do
         return if stopped?
 
+        # This is a safety check to ensure we haven't leaked
+        # processors somehow.
+        raise "BUG: No processors, cannot continue!" if @ready.empty? && @busy.empty?
+
         # Dispatch loop
         loop do
           break logger.debug('no processors') if @ready.empty?
@@ -117,7 +123,7 @@ module Sidekiq
           @ready.size.times do
             found ||= find_work(@queues.sample)
           end
-          break logger.debug('nothing to process') unless found
+          break unless found
         end
 
         # This is the polling loop that ensures we check Redis every