+++ myapp/app/workers/hard_worker.rb
@@ -2,6 +2,7 @@ class HardWorker
   include Sidekiq::Worker
 
   def perform(name, count, salt)
+    raise name if name == 'crash'
     print "#{Time.now}\n"
     sleep count
   end