+++ myapp/config/routes.rb
@@ -1,9 +1,10 @@
-require 'resque/server'
+require 'sidekiq/web'
 
 Myapp::Application.routes.draw do
-  mount Resque::Server.new, :at => '/resque'
+  mount Sidekiq::Web => '/sidekiq'
   get "work" => "work#index"
   get "work/email" => "work#email"
   get "work/post" => "work#delayed_post"
   get "work/long" => "work#long"
+  get "work/crash" => "work#crash"
 end