+++ test/helper.rb
@@ -1,5 +1,4 @@
-$TESTING = true
-if false
+if ENV.has_key?("SIMPLECOV")
   require 'simplecov'
   SimpleCov.start
 end
@@ -8,5 +7,9 @@ require 'minitest/unit'
 require 'minitest/pride'
 require 'minitest/autorun'
 
+require 'sidekiq'
 require 'sidekiq/util'
 Sidekiq::Util.logger.level = Logger::ERROR
+
+require 'sidekiq/redis_connection'
+REDIS = Sidekiq::RedisConnection.create(:url => "redis://localhost/15")