+++ test/test_cli.rb
@@ -2,10 +2,20 @@ require 'helper'
 require 'sidekiq/cli'
 require 'tempfile'
 
+cli = Sidekiq::CLI.instance
+def cli.die(code)
+  @code = code
+end
+
+def cli.valid?
+  !@code
+end
+
 class TestCli < MiniTest::Unit::TestCase
   describe 'with cli' do
+
     before do
-      @cli = new_cli
+      @cli = Sidekiq::CLI.instance
     end
 
     it 'blows up with an invalid require' do
@@ -14,7 +24,7 @@ class TestCli < MiniTest::Unit::TestCase
       end
     end
 
-    it 'blows up with invalid Ruby' do
+    it 'requires the specified Ruby code' do
       @cli.parse(['sidekiq', '-r', './test/fake_env.rb'])
       assert($LOADED_FEATURES.any? { |x| x =~ /fake_env/ })
       assert @cli.valid?
@@ -30,9 +40,22 @@ class TestCli < MiniTest::Unit::TestCase
       assert_equal ['foo'], Sidekiq.options[:queues]
     end
 
+    it 'changes timeout' do
+      @cli.parse(['sidekiq', '-t', '30', '-r', './test/fake_env.rb'])
+      assert_equal 30, Sidekiq.options[:timeout]
+    end
+
     it 'handles weights' do
       @cli.parse(['sidekiq', '-q', 'foo,3', '-q', 'bar', '-r', './test/fake_env.rb'])
       assert_equal %w(bar foo foo foo), Sidekiq.options[:queues].sort
+    end
+
+    it 'sets verbose' do
+      old = Sidekiq::Util.logger.level
+      @cli.parse(['sidekiq', '-v', '-r', './test/fake_env.rb'])
+      assert_equal Logger::DEBUG, Sidekiq::Util.logger.level
+      # If we leave the logger at DEBUG it'll add a lot of noise to the test output
+      Sidekiq::Util.logger.level = old
     end
 
     describe 'with pidfile' do
@@ -140,17 +163,6 @@ class TestCli < MiniTest::Unit::TestCase
         assert_equal 3, Sidekiq.options[:queues].select{ |q| q == 'seldom' }.length
       end
     end
-
-    def new_cli
-      cli = Sidekiq::CLI.new
-      def cli.die(code)
-        @code = code
-      end
-
-      def cli.valid?
-        !@code
-      end
-      cli
-    end
   end
+
 end