+++ test/test_client.rb
@@ -5,40 +5,42 @@ require 'sidekiq/worker'
 class TestClient < MiniTest::Unit::TestCase
   describe 'with real redis' do
     before do
-      Sidekiq.redis = { :url => 'redis://localhost/sidekiq_test' }
-      Sidekiq.redis.flushdb
+      Sidekiq.redis = REDIS
+      Sidekiq.redis {|c| c.flushdb }
     end
 
     it 'does not push duplicate messages when configured for unique only' do
       Sidekiq.client_middleware.entries.clear
       Sidekiq.client_middleware do |chain|
         chain.add Sidekiq::Middleware::Client::UniqueJobs
-        chain.add Sidekiq::Middleware::Client::ResqueWebCompatibility
       end
       10.times { Sidekiq::Client.push('customqueue', 'class' => 'Foo', 'args' => [1, 2]) }
-      assert_equal 1, Sidekiq.redis.llen("queue:customqueue")
+      assert_equal 1, Sidekiq.redis {|c| c.llen("queue:customqueue") }
     end
 
     it 'does push duplicate messages when not configured for unique only' do
       Sidekiq.client_middleware.remove(Sidekiq::Middleware::Client::UniqueJobs)
       10.times { Sidekiq::Client.push('customqueue2', 'class' => 'Foo', 'args' => [1, 2]) }
-      assert_equal 10, Sidekiq.redis.llen("queue:customqueue2")
+      assert_equal 10, Sidekiq.redis {|c| c.llen("queue:customqueue2") }
     end
   end
 
   describe 'with mock redis' do
     before do
       @redis = MiniTest::Mock.new
-      def @redis.multi; yield; end
+      def @redis.multi; yield if block_given?; end
       def @redis.set(*); true; end
       def @redis.sadd(*); true; end
       def @redis.srem(*); true; end
       def @redis.get(*); nil; end
       def @redis.del(*); nil; end
       def @redis.incrby(*); nil; end
-      def @redis.setex(*); nil; end
+      def @redis.setex(*); true; end
       def @redis.expire(*); true; end
+      def @redis.watch(*); true; end
       def @redis.with_connection; yield self; end
+      def @redis.with; yield self; end
+      def @redis.exec; true; end
       Sidekiq.instance_variable_set(:@redis, @redis)
     end