+++ test/test_processor.rb
@@ -8,6 +8,7 @@ class TestProcessor < MiniTest::Unit::TestCase
       $errors = []
       @boss = MiniTest::Mock.new
       Celluloid.logger = nil
+      Sidekiq.redis = REDIS
     end
 
     class MockWorker
@@ -27,26 +28,5 @@ class TestProcessor < MiniTest::Unit::TestCase
       assert_equal 1, $invokes
       assert_equal 0, $errors.size
     end
-
-    it 'handles exceptions' do
-      msg = { 'class' => MockWorker.to_s, 'args' => ['boom'] }
-      processor = ::Sidekiq::Processor.new(@boss)
-      assert_raises RuntimeError do
-        processor.process(msg, 'default')
-      end
-      @boss.verify
-      assert_equal 0, $invokes
-      assert_equal 1, $errors.size
-      assert_equal msg, $errors[0][:parameters]
-    end
-
   end
 end
-
-class FakeAirbrake
-  def self.notify(ex, hash)
-    $errors << hash
-  end
-end
-Airbrake = FakeAirbrake
-